+++ src/parse-query/index.ts
@@ -43,9 +43,12 @@ function collectKeywordedQueries(
     newQuery = newQuery.replace(group, "");
 
     const operator = (group.match(OPERATOR_REGEX) || [":"])[0];
-    const pair = group.split(operator);
-    const key = pair[0]?.toLowerCase().replace(/_/g, "");
-    let value = pair[1];
+    const indexOfOperator = group.indexOf(operator);
+    const key = group
+      .substring(0, indexOfOperator)
+      .toLowerCase()
+      .replace(/_/g, "");
+    let value = group.substring(indexOfOperator + 1, group.length);
 
     if (value.length > 2) {
       const firstChar = value.charAt(0);