+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# unreleased
+
+`search`
+  - Fix bug where `:` could not be used in search terms
+
 # 0.13.0
 
 - `search`